+++ src/algorithms/tracking/gnuradio_blocks/dll_pll_veml_tracking_fpga.cc
@@ -1251,14 +1251,14 @@ int dll_pll_veml_tracking_fpga::general_work (int noutput_items __attribute__((u
             {
                 d_pull_in = 0;
                 multicorrelator_fpga->lock_channel();
-                unsigned counter_value = multicorrelator_fpga->read_sample_counter();
+                unsigned long int counter_value = multicorrelator_fpga->read_sample_counter();
                 //printf("333333 counter_value = %d\n", counter_value);
                 //printf("333333 current_synchro_data.Acq_samplestamp_samples = %d\n", current_synchro_data.Acq_samplestamp_samples);
                 //printf("333333 current_synchro_data.Acq_delay_samples = %f\n", current_synchro_data.Acq_delay_samples);
                 //printf("333333 d_correlation_length_samples = %d\n", d_correlation_length_samples);
                 unsigned num_frames = ceil((counter_value - current_synchro_data.Acq_samplestamp_samples - current_synchro_data.Acq_delay_samples)/d_correlation_length_samples);
                 //printf("333333 num_frames = %d\n", num_frames);
-                unsigned absolute_samples_offset = current_synchro_data.Acq_delay_samples + current_synchro_data.Acq_samplestamp_samples + num_frames*d_correlation_length_samples;
+                unsigned long int absolute_samples_offset = current_synchro_data.Acq_delay_samples + current_synchro_data.Acq_samplestamp_samples + num_frames*d_correlation_length_samples;
                 //printf("333333 absolute_samples_offset = %d\n", absolute_samples_offset);
                 multicorrelator_fpga->set_initial_sample(absolute_samples_offset);
                 d_absolute_samples_offset = absolute_samples_offset;
